Wonderful: libgcc: arm: optimize __muldi3 and __mulvsi3 on Thumb1 architectures
Patch the __muldi3 and __mulvsi3 functions to be compiled using an ARM
mode function for ARMv4/ARMv5/ARMv6 in Thumb mode. In those two specific cases,
the ARM mode code is substnatially smaller and faster than the Thumb mode code;
a precedent has already been made with __clzsi2. In addition, the code for
__muldi3 has manually been adjusted in assembly to avoid an unnecessary mov.